Added dashboard menu
