+++ client/src/app/globals.css
@@ -15,3 +15,36 @@
     --background-end-rgb: 0, 0, 0;
   }
 }
+
+.menuItem {
+  position: relative;
+  @apply bg-blue-300 w-72 h-72 p-3 m-5 rounded-md flex justify-center items-center text-white text-lg;
+  background-color: #009be5;
+}
+.menuItem::before,
+.menuItem::after {
+  content: "";
+  position: absolute;
+  top: 0;
+  bottom: 0;
+  width: 2px;
+  background-color: #fc2f70;
+  transform: scaleY(0);
+  transition: transform 0.5s ease;
+}
+.menuItem::before {
+  left: -8px;
+  transform-origin: center top;
+}
+.menuItem:hover::before {
+  transform-origin: center bottom;
+  transform: scaleY(1);
+}
+.menuItem::after {
+  right: -8px;
+  transform-origin: center bottom;
+}
+.menuItem:hover::after {
+  transform-origin: center top;
+  transform: scaleY(1);
+}
