+++ client/src/auth.config.ts
@@ -9,12 +9,12 @@ export const authConfig = {
       return true;
       const isLoggedIn = !!auth?.user;
       console.log({ isLoggedIn, auth });
-      const isOnDashboard = nextUrl.pathname.startsWith("/");
+      const isOnDashboard = nextUrl.pathname.startsWith("/dashboard");
       if (isOnDashboard) {
         if (isLoggedIn) return true;
         return false; // Redirect unauthenticated users to login page
       } else if (isLoggedIn) {
-        return Response.redirect(new URL("/", nextUrl));
+        return Response.redirect(new URL("/dashboard", nextUrl));
       }
       return true;
     },
