+++ client/src/auth.ts
@@ -3,7 +3,7 @@ import Credentials from "next-auth/providers/credentials";
 import { authConfig } from "./auth.config";
 import { z } from "zod";
 import { sql } from "@vercel/postgres";
-import type { User } from "./app/lib/definition";
+import type { User } from "./_lib/definition";
 import bcrypt from "bcryptjs-react";
 
 async function getUser(email: string): Promise<User | undefined> {
